feat: add RabbitMQ producer for notification messages
This commit adds a new RabbitMQ producer for sending notification messages. The producer connects to the RabbitMQ server, creates an exchange, a queue, and binds the queue to the exchange. It then sends a notification message to the queue. The message contains information about a new product being added. The producer also sets an expiration time for the message.